// src/components/Theme/types.ts
export type ThemeSimpleBorderRadius = 'none' | 'base';
export type ThemeBorderRadius =
  | ThemeSimpleBorderRadius
  | 'fullyRounded'
  | 'tight';
export type ThemeBorderColor = 'base' | 'emphasized';
export type ThemeSimpleBorder = 'none' | 'full';
export type ThemeBorder = ThemeSimpleBorder | 'blockEnd';
export type ThemeResourceListBorder =
  | ThemeBorder
  | 'betweenItems'
  | 'aroundItems';
export type ThemeBorderStyle = 'base' | 'dotted';
export type ThemeBorderWidth = 'base' | 'thick';

export type ThemeSpacing =
  | 'none'
  | 'tight4x'
  | 'tight3x'
  | 'tight2x'
  | 'tight1x'
  | 'tight'
  | 'base'
  | 'loose'
  | 'loose1x'
  | 'loose2x'
  | 'loose3x'
  | 'loose4x';
// ...
export type ThemeBackground =
  | 'surfacePrimary'
  | 'surfaceSecondary'
  | 'surfaceTertiary'
  | 'transparent';